pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rl_commons"
version = "0.1.0"
description = "A settings generator from YAML to Python dataclass"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "ruamel.yaml",
  "load-dotenv",
  "dacite",
  "coloredlogs"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
generate-settings = "settingsgen.generate_settings:main"